3°semestre/Implementação orientada a objetos(JAVA)/Programas/Exemplos/src/Aluno.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

/**
 *
 * @author alunocmc
 */
public class Aluno {
private int rgm;
private String nome;
private double nota1;
private double nota2;
private double notaFinal;

private boolean situacao;
    public double getNotaFinal() {
        return notaFinal;
    }

    public void setNotaFinal(double notaFinal) {
        this.notaFinal = notaFinal;
    }

    public int getRgm() {
        return rgm;
    }

    public void setRgm(int rgm) {
        this.rgm = rgm;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public double getNota1() {
        return nota1;
    }

    public void setNota1(double nota1) {
        this.nota1 = nota1;
    }

    public double getNota2() {
        return nota2;
    }

    public void setNota2(double nota2) {
        this.nota2 = nota2;
    }
    public void calcularMedia(){
        notaFinal=(nota1+nota2*2)/3;
    }
    public boolean verificarSituacao(){
        if(notaFinal>=5){
            situacao = true;
            return situacao;
        }else{
            situacao=false;
            return situacao;
        }
    }
}
